Move after-request header setting to views/base.py from views/core.py
